dogfooding changes for dataset/team objects
