+++ darwin/future/meta/objects/team.py
@@ -1,10 +1,13 @@
-from typing import List, Optional
+from typing import List, Optional, Tuple, Union
 
 from darwin.future.core.client import ClientCore
+from darwin.future.core.datasets import get_dataset, remove_dataset
 from darwin.future.core.team.get_team import get_team
 from darwin.future.data_objects.team import TeamCore
+from darwin.future.exceptions import MissingDataset
 from darwin.future.helpers.assertion import assert_is
 from darwin.future.meta.objects.base import MetaBase
+from darwin.future.meta.objects.dataset import Dataset
 from darwin.future.meta.queries.dataset import DatasetQuery
 from darwin.future.meta.queries.team_member import TeamMemberQuery
 from darwin.future.meta.queries.workflow import WorkflowQuery
@@ -55,3 +58,89 @@ class Team(MetaBase[TeamCore]):
 
     def __str__(self) -> str:
         return f"TeamMeta(name='{self.name}', slug='{self.slug}', id='{self.id}' - {len(self._element.members if self._element.members else [])} members)"
+
+    @classmethod
+    def delete_dataset(cls, client: ClientCore, dataset_id: Union[int, str]) -> Tuple[Optional[List[Exception]], int]:
+        """
+        Deletes a dataset by id or slug
+
+        Parameters
+        ----------
+        dataset_id: Union[int, str]
+            The id or slug of the dataset to delete
+
+        Returns
+        -------
+        Tuple[Optional[List[Exception]], int]
+            A tuple containing a list of exceptions and the number of datasets deleted
+        """
+        exceptions = []
+        dataset_deleted = -1
+
+        try:
+            if isinstance(dataset_id, str):
+                dataset_deleted = cls._delete_dataset_by_slug(client, dataset_id)
+            else:
+                dataset_deleted = cls._delete_dataset_by_id(client, dataset_id)
+
+        except Exception as e:
+            exceptions.append(e)
+
+        return exceptions or None, dataset_deleted
+
+    @staticmethod
+    def _delete_dataset_by_slug(client: ClientCore, slug: str) -> int:
+        """
+        (internal) Deletes a dataset by slug
+
+        Parameters
+        ----------
+        client: MetaClient
+            The client to use to make the request
+
+        slug: str
+            The slug of the dataset to delete
+
+        Returns
+        -------
+        int
+            The dataset deleted
+        """
+        assert_is(isinstance(client, ClientCore), "client must be a Core Client")
+        assert_is(isinstance(slug, str), "slug must be a string")
+
+        dataset = get_dataset(client, slug)
+        if dataset and dataset.id:
+            dataset_deleted = remove_dataset(client, dataset.id)
+        else:
+            raise MissingDataset(f"Dataset with slug {slug} not found")
+
+        return dataset_deleted
+
+    @staticmethod
+    def _delete_dataset_by_id(client: ClientCore, dataset_id: int) -> int:
+        """
+        (internal) Deletes a dataset by id
+
+        Parameters
+        ----------
+        client: Client
+            The client to use to make the request
+
+        dataset_id: int
+            The id of the dataset to delete
+
+        Returns
+        -------
+        int
+            The dataset deleted
+        """
+        assert_is(isinstance(client, ClientCore), "client must be a Client")
+        assert_is(isinstance(dataset_id, int), "dataset_id must be an integer")
+
+        dataset_deleted = remove_dataset(client, dataset_id)
+        return dataset_deleted
+
+    def create_dataset(self, slug: str) -> Dataset:
+        core = Dataset.create_dataset(self.client, slug)
+        return Dataset(self.client, core, meta_params={"team_slug": self.slug})
