+++ darwin/future/tests/meta/objects/fixtures.py
@@ -4,9 +4,11 @@ from uuid import UUID
 from pytest import fixture, raises
 
 from darwin.future.core.client import ClientCore
+from darwin.future.data_objects.dataset import DatasetCore
 from darwin.future.data_objects.team import TeamCore
 from darwin.future.data_objects.workflow import WFStageCore, WorkflowCore
 from darwin.future.meta.objects import stage
+from darwin.future.meta.objects.dataset import Dataset
 from darwin.future.meta.objects.stage import Stage
 from darwin.future.meta.objects.team import Team
 from darwin.future.meta.objects.workflow import Workflow
@@ -36,3 +38,8 @@ def base_meta_stage(base_client: ClientCore, base_stage: WFStageCore, base_UUID:
 @fixture
 def base_meta_stage_list(base_meta_stage: Stage, base_UUID: UUID) -> List[Stage]:
     return [base_meta_stage]
+
+
+@fixture
+def base_meta_dataset(base_client: ClientCore, base_dataset: DatasetCore) -> Dataset:
+    return Dataset(base_client, base_dataset, meta_params={"team_slug": "test_team"})
